+++ src/enums/VariableType.ts
@@ -1,9 +1,0 @@
-enum VariableType {
-  String = 'String',
-  Number = 'Number',
-  Boolean = 'Boolean',
-  Date = 'Date',
-  Address = 'Address',
-}
-
-export default VariableType;